backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
python-pptx==1.0.2
pydantic==2.9.0
